requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
boto3==1.26.84
botocore==1.29.84
Django==4.1.7
django-storages==1.13.2
jmespath==1.0.1
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-decouple==3.8
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
